Configure to review nan values and start modeling
First review the data to work with the empty data
and clean this values. Input the na values of Age
and drop some rows and columns that not apport
the analysis.

Finally, start the models, with the linear regression.
Althought, we must change because we want to predict
the category, and for this, I have to use the Logistic
regression.