Change many of the develop to improve the project
I have to configure the version of python to 3.11.7
because some modules are not allows in the version 3.12.
Also, install tensorflow and pycaret, to implement ML
models.

Start with the pycaret package and found the best classifier
model. With this, we use to predict and obtain the results in the
challenge.